Partway through refactoring of Arena.
